feat:抽取 Controller 共用部分
