+++ controllers/pages/restaurant-controller.js
@@ -1,45 +1,9 @@
 const { Restaurant, Category, Comment, User, Favorite } = require('../../models')
-const { getOffset, getPagination } = require('../../helpers/pagination-helper') // 加入這行
+const restaurantServices = require('../../services/restaurant-services') // 引入 restaurant-services
 const favorite = require('../../models/favorite')
 const restaurantController = {
-  getRestaurants: (req, res, next) => { // 補上 next
-    // 修改以下
-    const DEFAULT_LIMIT = 9 // 加入這行
-    const categoryId = Number(req.query.categoryId) || '' // 新增這裡，從網址上拿下來的參數是字串，先轉成 Number 再操作
-    // 新增以下
-    const page = Number(req.query.page) || 1
-    const limit = Number(req.query.limit) || DEFAULT_LIMIT
-    const offset = getOffset(limit, page) // 增加這裡
-    return Promise.all([
-      Restaurant.findAndCountAll({
-        include: Category,
-        where: { // 新增查詢條件
-          ...categoryId ? { categoryId } : {} // 檢查 categoryId 是否為空值
-        },
-        limit, // 增加這裡
-        offset, // 增加這裡
-        nest: true,
-        raw: true
-      }),
-      Category.findAll({ raw: true })
-    ])
-      .then(([restaurants, categories]) => {
-        const favoritedRestaurantsId = req.user && req.user.FavoritedRestaurants.map(fr => fr.id) // 新增這一行
-        const likedRestaurantsId = req.user && req.user.LikedRestaurants.map(lk => lk.id) // 新增這一行
-        const data = restaurants.rows.map(r => ({ // 修改這裡，加上 .rows
-          ...r,
-          description: r.description.substring(0, 50),
-          isFavorited: favoritedRestaurantsId.includes(r.id), // 修改這一行
-          isLiked: likedRestaurantsId.includes(r.id) // 修改這一行
-        }))
-        return res.render('restaurants', {
-          restaurants: data,
-          categories,
-          categoryId, // 新增這裡
-          pagination: getPagination(limit, page, restaurants.count) // 修改這裡，把 pagination 資料傳回樣板
-        })
-      })
-      .catch(err => next(err)) // 補上
+  getRestaurants: (req, res, next) => {
+    restaurantServices.getRestaurants(req, (err, data) => err ? next(err) : res.render('restaurants', data))
   },
   getRestaurant: (req, res, next) => {
     return Restaurant.findByPk(req.params.id, {
